Fixed data sync issue
