+++ app/src/main/res/layout/recycler_list_row.xml
@@ -58,7 +58,6 @@
                 android:id="@+id/from"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:ellipsize="end"
                 android:lines="1"
                 android:textColor="@color/from"
                 android:textSize="@dimen/msg_text_primary"
@@ -68,7 +67,6 @@
                 android:id="@+id/txt_primary"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:ellipsize="end"
                 android:lines="1"
                 android:textColor="@color/subject"
                 android:textSize="@dimen/msg_text_secondary"
@@ -78,7 +76,6 @@
                 android:id="@+id/txt_secondary"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:ellipsize="end"
                 android:lines="1"
                 android:textColor="@color/message"
                 android:textSize="@dimen/msg_text_secondary" />
